Feature music

+++ src/main/java/com/wip/bool/music/mp3/service/SongMP3Service.java
@@ -1,11 +1,16 @@
 package com.wip.bool.music.mp3.service;
 
+import com.wip.bool.exception.excp.AuthorizationException;
 import com.wip.bool.exception.excp.EntityNotFoundException;
 import com.wip.bool.exception.excp.ErrorCode;
+import com.wip.bool.exception.excp.FileException;
 import com.wip.bool.music.mp3.domain.SongMP3;
 import com.wip.bool.music.mp3.domain.SongMP3Repository;
 import com.wip.bool.music.song.domain.SongDetail;
 import com.wip.bool.music.song.domain.SongDetailRepository;
+import com.wip.bool.user.domain.Role;
+import com.wip.bool.user.domain.User;
+import com.wip.bool.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -19,44 +24,64 @@ public class SongMP3Service {
     private String mp3FilePath;
 
     private final SongDetailRepository songDetailRepository;
+
     private final SongMP3Repository songMP3Repository;
 
+    private final UserRepository userRepository;
+
     @Transactional
-    public Long save(Long songDetailId, String orgFileName, byte[] mp3File) {
+    public Long saveSongMP3(Long userId, Long songDetailId, String orgFileName, byte[] mp3File) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException(userId, ErrorCode.NOT_FOUND_USER));
+
+        Role role = user.getRole();
+
+        if(role != Role.ROLE_ADMIN) {
+            throw new AuthorizationException();
+        }
 
         SongDetail songDetail = songDetailRepository.findById(songDetailId)
                 .orElseThrow(() -> new EntityNotFoundException(songDetailId, ErrorCode.NOT_FOUND_SONG));
 
         SongMP3 songMP3 = SongMP3.createSongMP3(songDetail, mp3FilePath, orgFileName, mp3File);
 
-        songMP3Repository.save(songMP3);
+        Long id = songMP3Repository.save(songMP3).getId();
 
         if(!songMP3.createMP3File(mp3FilePath)) {
-            throw new IllegalStateException("mp3 파일 생성이 실패했습니다.");
+            throw new FileException(ErrorCode.CREATE_FAIL_FILE);
         }
 
         if(!songMP3.updateMP3Info(mp3FilePath)) {
             songMP3.deleteMP3File(mp3FilePath);
-
-            throw new IllegalStateException("mp3 파일 생성이 실패했습니다.");
+            throw new FileException(ErrorCode.CREATE_FAIL_FILE);
         }
 
-        return songMP3.getId();
+        return id;
     }
 
     @Transactional
-    public Long delete(Long songMP3Id) {
+    public Long deleteSongMP3(Long userId, Long songMP3Id) {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException(userId, ErrorCode.NOT_FOUND_USER));
+
+        Role role = user.getRole();
+
+        if(role != Role.ROLE_ADMIN) {
+            throw new AuthorizationException();
+        }
 
         SongMP3 songMP3 = songMP3Repository.findById(songMP3Id)
             .orElseThrow(() -> new EntityNotFoundException(songMP3Id, ErrorCode.NOT_FOUND_MP3));
 
-        songMP3Repository.delete(songMP3);
+        Long resValue = songMP3Repository.delete(songMP3);
 
         if(!songMP3.deleteMP3File(mp3FilePath)) {
-            throw new IllegalStateException("mp3 파일 삭제가 실패했습니다.");
+            throw new FileException(ErrorCode.DELETE_FAIL_FILE);
         }
 
-        return 1L;
+        return resValue;
     }
 
     @Transactional(readOnly = true)
@@ -65,7 +90,7 @@ public class SongMP3Service {
         SongMP3 songMP3 = songMP3Repository.findById(songMP3Id)
                 .orElseThrow(() -> new EntityNotFoundException(songMP3Id, ErrorCode.NOT_FOUND_MP3));
 
-        return songMP3.getFile(mp3FilePath);
+        return songMP3.getFile();
     }
 
 }
